requirements.txt
torch>=1.9.0
torchvision>=0.10.0
rdkit-pypi>=2021.9.4
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
transformers>=4.12.0
biopython>=1.79
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.4.0